requirements.txt
pydantic
streamlit
python-dotenv
pandas
agno
exa-py
google-generativeai
fastapi
uvicorn
